pyproject.toml
[tool.poetry]
name = "aws-cloud-project"
version = "0.1.0"
description = ""
authors = ["Rafael Panegassi <rm.panegassi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.13.0"
boto3 = "^1.38.3"
python-dotenv = "^1.1.0"
loguru = "^0.7.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
